Add increment-part-power command to swift-ring-builder.
It increases the part_power of the given ring by 1 without causing any
data to move. If your cluster has gotten too big and you don't have
enough partitions per device, this lets you get more partitions to
level things out without causing a huge I/O storm. The old method
(construct a completely new ring with a higher part_power) can cause
most of your objects to move, which can completely destroy the
performance and availability of your cluster.

Since this is an irreversible operation, a backup of the builder file
is made, just like rebalance does.

Having this operation available makes setting up a Swift cluster a lot
less scary. If you're a sysadmin setting up a Swift cluster for the
first time, you've got to pick a value for part_power that's big
enough to accommodate all the future growth you ever think you'll
have, but not so big that your Swift daemons chew up all your RAM and
drive your boxes into swap. Getting it wrong used to mean great pain
in your future; with this commit, though, the sysadmin can start with
a smallish part_power and increase it as needed.

Here's how it works:

Let's say you have a ring with part_power 3. The part2dev mapping
might look like this:

000: 0
001: 2
010: 1
011: 1
100: 2
101: 0
110: 0
111: 1

Remember that, to figure out which device an object lives on, we take
the first part_power bits of the hash of the object's name (I'm
ignoring the stuff about replicas for simplicity). If we have an
object named "kitten.jpg", the md5sum of that is a42359fc..., and the
first 3 bits are 101, and hence kitten.jpg will live on device 0.

So, to increment the part_power without moving any partitions around,
we have to construct a new part2dev mapping that's twice as big but
still ends up mapping every checksum to the same place as before. We
can construct the new part2dev by simply repeating each entry in the
old part2dev twice, like so:

0000: 0
0001: 0
0010: 2
0011: 2
0100: 1
0101: 1
0110: 1
0111: 1
1000: 2
1001: 2
1010: 0
1011: 0
1100: 0
1101: 0
1110: 1
1111: 1

We look at the first 4 bits of the md5 of "kitten.jpg" to get 1010,
which maps to device 0, same as before.

In general, if your hash mapped to index N before the increment of
part_power, it maps to either index 2N or 2N+1 after the
increment. Since each entry in the old map is repeated twice, the
value at the new map's indices 2N and 2N+1 is the same as the value at
the old map's index N.