Add general multi-ring support
The basic idea here is to replace the use of a single object ring in the Application
class with a collection of object rings. The collection includes not only the Ring
object itself but the policy name assocaited with it, the filename for the .gz and
any other meta data associated with the policy that may be needed.

The policy collection is based off of info in the swift.conf file parsed at wsgi
application init time. The format of the sections in the .conf file is covered
below and accessed via get_storage_policies() that has all of the info except
for the Ring addresses; the rings are instantiated when Application() is instantiated
and then Application.object.nodes is updated with the addresses. There is a new
accessor method called Application.get_object_ring() that returns the Ring address
based on the policy passed in. There is basic error handling in both the conf file
parser as well as the accessor functions to assure the conf file is formed correctly.

swift.conf format:

	[storage-policy:0]
	name = chicken

	[storage-policy:1]
	name = turkey
	default = yes

	[storage-policy:2]
	name = gravy
	ec-scheme = 14:10

With the above format:

- policy 0 will always be used for access to existing containers w/o the policy specified.
The ring name for polic 0 is always 'object' assuring backwards compatiblity.  The parser
will always create a policy 0 even if not specified
- the policy with 'default=yes' is the one used for new container creation and allows the
admin to specify which policy is used without forcing the application to add the metadata.
- use of 'ec-scheme' implies that the associated ring is built for erasure code and
includes the ec scheme parameter.  A policy w/o any other tag like this implies replication
using where the # of copies is determined by how the ring was built

This commit focuses only on the Proxy server (storage node services using the obj
rings will be next) and the unit test framework. No new tests have been added yet
as the number of changes required to support multiple rings in the test framwork
was large.

Ring selection is done at public entry points based on container metadata. If the
container has the metadata X-Container-Storage-Policy and there is a match
with the policy name provided and one parsed from swift.conf, that ring will be
used. If not then decisions are made per description of swift.conf sectiosn above.

The container metadata policy cannot be changed, if an attempt is made to change it a c
onflict error will be returned. An attempt to create a container with a policy not defined
results in a bad reqeust error.  Migrating data from one policy to another means copying it
from one container to another manually.

The following highlights the major changes in this commit:

* swift.conf file parser support for storage policies
* container update to support immutable storage policy metadata
* a few misc function parm changes to enable passing of ring info
* use of ring based on container metadata in objController public functions
* large number of changes to unit test framework to support multiple rings
but no new capabilities specifically geared at testing multiple rings

Change-Id: I53bcad1dc724596c5c4ec8a30d5cf50ebe7f1b7d
Implements:  blueprint ec-proxy-work