Only cache mount checks on read-only operations.
If we write an object to the root FS on a PUT, or write its tombstone
to the root FS on a DELETE, or something similar, then that data is
lost as soon as the drive is re-mounted. That's not good for
durability; we can have an object PUT that gets written to only 1 of 3
storage drives; one connection could die in the middle, and one copy
could get written to a root FS and then be mounted over.

Change-Id: I8ea78380cfb8fed84ce0088e8618c24e3bf2a0da