Make hash algorithm configurable for new rings.
While it is true that there are currently no second-preimage attacks
against MD5, there are other attacks that weaken it, and it's not like
MD5 is going to get any stronger.

Now you can give a different hash algorithm to `swift-ring-builder
create` on the command line, and all your object paths will get hashed
with the new algorithm. On-disk filenames will reflect the new hash.

You can't change an existing ring's hash algorithm. I mean, you *can*,
but you'll lose all your data.

This works by having the proxy (a) pay attention to the ring's hash
algorithm when hashing paths, and (b) extract the hash algorithm from
the rings and pass it to the backend servers via HTTP headers so they
can hash paths the same way. If the header is missing, the backends
will default to md5, so upgrades of existing clusters will work.

Note that there's a use of MD5 in Ring.get_more_nodes() that's been
left there intentionally. It doesn't affect on-disk placement, only
handoff selection, and it operates on the partition number that comes
from the configurable hash anyway. Changing that wouldn't actually
gain us anything.

Note: doesn't work on py2.6, it seems. Who knew?

Change-Id: I48eb975ac3c252294f365d19f7aaff96c507b468