Rework the interface to storage policies just a little.
The main thing was to make the code look a little more like the rest
of the Swift codebase; things like using full words ("storage" instead
of "stor", for example) and being a little more testable.

In particular, controllers now take an optional
StoragePolicyCollection in their constructors, so now there's no
checks for an index of "unit-test" in the code. It wasn't too bad, but
it's sort of gross to have a different code path for tests in the real
code. I've been burned by that before on previous projects.

Still to do: figure out what data type an index is. int? str? it
should be one everywhere.

Still to do: simplify the plumbing around getting the storage policy
thingy out of the headers; we can raise HTTPBadRequest instead of
checking return values. That's how a bunch of other request-validation
code works, so this should try to blend in.

Still to do: pass the index to the container server, not the policy
name. Since an operator *could* change names, we need to send policy
*indices* so object async_updates don't get invalidated by operator
stupidity.

Change-Id: I4f0b12ce3b8c8e12020f09a1dbaeac7174125c1d