Store policy index in container_stat table
The eventual goal is to have policy index stored per-object in the
objects table and have a count of mismatched objects in the
container_stat table which is magically kept up-to-date by a
trigger. As part of that, I need the trigger to know the container's
policy index, so I moved it to a convenient place.

I considered making an extract_spi sqlite function similiar to chexor
and calling that from the trigger, which would go something like this:

 * read a multi-kilobyte blob of text out of container_stat
 * pass it from sqlite to python
 * run it through a JSON decoder
 * pull out one single value
 * return that from python back into sqlite

Of course, once I saw the data flow there I ran away screaming and put
policy index in container_stat instead.

Change-Id: I26f7ef09201970ee876ac24f39d0b2d4895bfda9