Fix changing of storage policy index
The commit that the storage policy moved out of metadata and into
container_stat introduced a bug: if the client DELETEs and then PUTs a
container with a different storage policy, the resurrected container
still had the old policy. This is because the policy was being set in
broker.initialize(), but that was only being called for new
containers (no .db files at all).

Now the policy index is still set at initialization for new
containers, but it is updated for resurrected containers.

I considered using broker.set_storage_policy_index() for all
containers, new or resurrected, but that would leave a brief window of
time for new containers where the .db file has the default policy
index and not the user-specified one, and that seemed like a good home
for bugs to move into.

Change-Id: I496e5078df80ec1466c1d50ac62ae3700c5aeec0