Add container reconciliation to container updater
Under normal circumstances, an object lives in storage policy N, where
N is the value from its container's container_stat table. However,
under some circumstances involving network failures and such, it can
happen that an object is actually stored in a different policy than
its container.

This commit introduces functionality in the container updater to move
misplaced objects to their proper places. I chose the container
updater because it's already doing a filesystem walk and inspecting
every container DB, so when nothing is misplaced, no extra work is
being done. I used a separate greenpool for the work so that account
stats would still be updated with the same latency as before.

******************** NOT DONE ****************************

stuff that's not done

 * still need statsd tests (marked w/XXX comments)

 * really really needs a probetest or two

 * benchmark ContainerBroker.put_object and make sure it's not too much slower when SPIs match

Change-Id: I92ea9ddc80d89b21c7d98bdfde27eb221b9629ac