First pass at EC GET path
This handles the happy and slightly-less-happy paths for EC GET and
HEAD.

In the happy path, enough fragment archives are found to reconstitute
the object. Also, 404s in EC containers work.

Theoretically, EC GET will resume downloads like replicated GET if an
object server fails halfway through. That's not really tested yet,
though. There are tests for some of the lower-level functionality, but
there isn't one that kills an object server partway through.

The approach is basically Clay's approach with the two object
controllers. The proxy server picks the right one based on the
container's storage policy. The vast majority of the old
ObjectController is the same, but the class name changed to
BaseObjectController, and then there's a couple derived classes to
hold EC-specific and replication-specific behavior.

Change-Id: I988335aa1954d1e9d092c2eab3b3da9d89d1c19e