Add live status report to object replicator
The object replicator now has an internal scoreboard that contains
information about the currently-running replication jobs and also has
a socket interface for querying it.

The replicator's status report has things like partition, device,
storage policy, and peer to whom it is pushing for each active
replication job, plus the current stats for the replication pass and
the replicator's PID (to help with strace). It's easy to add more
things; anything JSON-serializable will work.

There's a new CLI tool called "swift-interrogate" that you can use to
query the running replicators. The output is pretty crude at the
moment, but it lets you see what your replicators are doing.

The socket interface uses Unix sockets in /var/run/swift, one per
replicator process, except for the master process in multi-process
mode as that process doesn't replicate anything.

The protocol is length-type-value; values are serialized JSON-RPC 2.0,
and the type is always 1. The type is there in case we outgrow
JSON-RPC.

Change-Id: I16f68dae9123838b87712635f05421d791a3efd8